alos reader fixes

+++ nest-reader-ers-ceos/src/test/java/org/esa/nest/dataio/ceos/records/FilePointerRecordTest.java
@@ -85,21 +85,21 @@ public class FilePointerRecordTest extends TestCase {
 
         assertNotNull(record);
         assertEquals("A ", record.getAttributeString("Ascii code character"));
-        assertEquals(2, record.getAttributeInt("File Pointer Number"));
+        assertEquals(2, (int)record.getAttributeInt("File Pointer Number"));
         assertEquals("AL PSMB2IMGYBSQ ", record.getAttributeString("File ID"));
         assertEquals("IMAGERY                     ", record.getAttributeString("File class"));
         assertEquals("IMGY", record.getAttributeString("File class code"));
         assertEquals("BINARY ONLY                 ", record.getAttributeString("File datatype"));
         assertEquals("BINO", record.getAttributeString("File datatype Code"));
-        assertEquals(14001, record.getAttributeInt("Number of records"));
-        assertEquals(897623, record.getAttributeInt("FirstRecordLength"));
-        assertEquals(8634264, record.getAttributeInt("MaxRecordLength"));
+        assertEquals(14001, (int)record.getAttributeInt("Number of records"));
+        assertEquals(897623, (int)record.getAttributeInt("FirstRecordLength"));
+        assertEquals(8634264, (int)record.getAttributeInt("MaxRecordLength"));
         assertEquals("FIXED LENGTH", record.getAttributeString("RecordLengthType"));
         assertEquals("FIXD", record.getAttributeString("RecordLengthTypeCode"));
-        assertEquals(1, record.getAttributeInt("FirstRecordVolumeNumber"));
-        assertEquals(2, record.getAttributeInt("FinalRecordVolumeNumber"));
-        assertEquals(3, record.getAttributeInt("ReferencedFilePortionStart"));
-        assertEquals(17, record.getAttributeInt("ReferencedFilePortionEnd"));
+        assertEquals(1, (int)record.getAttributeInt("FirstRecordVolumeNumber"));
+        assertEquals(2, (int)record.getAttributeInt("FinalRecordVolumeNumber"));
+        assertEquals(3, (int)record.getAttributeInt("ReferencedFilePortionStart"));
+        assertEquals(17, (int)record.getAttributeInt("ReferencedFilePortionEnd"));
     }
 
     private static void writeRecordData(ImageOutputStream ios) throws IOException {
